pyproject.toml
[project]
name = "LearningLangChain"
version = "0.1.0"
description = "Learning LangChain from CampusX"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "google-generativeai>=0.8.6",
    "grandalf>=0.8",
    "huggingface-hub[hf-xet]>=0.36.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.2.6",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]
